Move Merger and subclasses to merger module
